Fix for old browsers that don't support ArrayBufferViews in Blob constructor
Safari 6 and perhaps others supported the Blob constructor but not BlobBuilder. However, ArrayBufferViews in the Blob constructor were broken or not supported. Unfortunately, the tests didn't catch this since they weren't being run at all if Blob was undefined.

This separates out that check and maps ArrayBufferViews to ArrayBuffers in the Blob constructor as well as BlobBuilder when necessary.